fix: fix bug with page doubling while clicking next page
